Move BGP Policy test proto since //policytest/ depends on it.
Want to be able to exclude bgp/... without excluding valpb/...
