+++ integration_tests/dut_policy_tests/prefix_set/BUILD
@@ -9,9 +9,9 @@ go_test(
         "exclusive",
     ],
     deps = [
-        "//bgp/tests/proto/policyval",
         "//internal/binding",
         "//policytest",
+        "//proto/policyval",
         "@com_github_openconfig_ondatra//:ondatra",
         "@com_github_openconfig_ondatra//gnmi",
         "@com_github_openconfig_ondatra//gnmi/oc",
