+++ integration_tests/dut_policy_tests/prefix_set/prefix_set_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/openconfig/ondatra/gnmi"
 	"github.com/openconfig/ondatra/gnmi/oc"
 
-	valpb "github.com/openconfig/lemming/bgp/tests/proto/policyval"
+	valpb "github.com/openconfig/lemming/proto/policyval"
 )
 
 func TestMain(m *testing.M) {
